requirements.txt.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
langchain-ollama>=0.1.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
pypdf>=3.0.0
langchain-huggingface

